# Optimize reflection of F# types, part 2