Continuation of #9714.

- `PreComputeRecordConstructor`
- `PreComputeRecordFieldReader`
- `PreComputeRecordReader` - reworked from last time
- `PreComputeUnionConstructor`
- `PreComputeUnionReader`
- `PreComputeUnionTagReader`

`PreComputeTupleReader` and `PreComputeTupleConstructor` are a little more complicated, so perhaps another time.

The rest of the `PreCompute` family only return corresponding `System.Reflection.MethodBase` for further reflection use.

|                        Type |     Method |       Mean |      Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------- |----------- |-----------:|-----------:|----------:|-------:|------:|------:|----------:|
| PreComputeRecordConstructor | Reflection | 554.210 ns |  3.3056 ns | 3.0921 ns | 0.0134 |     - |     - |     112 B |
| PreComputeRecordConstructor |   Compiled |   9.999 ns |  0.0416 ns | 0.0389 ns | 0.0057 |     - |     - |      48 B |
|                             |            |            |            |           |        |       |       |           |
| PreComputeRecordFieldReader | Reflection | 112.941 ns |  1.9830 ns | 1.8549 ns | 0.0029 |     - |     - |      24 B |
| PreComputeRecordFieldReader |   Compiled |   3.431 ns |  0.1047 ns | 0.1120 ns | 0.0029 |     - |     - |      24 B |
|                             |            |            |            |           |        |       |       |           |
|      PreComputeRecordReader | Reflection | 562.094 ns | 10.2099 ns | 9.5503 ns | 0.0134 |     - |     - |     112 B |
|      PreComputeRecordReader |   Compiled |  22.212 ns |  0.2661 ns | 0.2489 ns | 0.0134 |     - |     - |     112 B |
|                             |            |            |            |           |        |       |       |           |
|  PreComputeUnionConstructor | Reflection | 473.935 ns |  5.1988 ns | 4.8629 ns | 0.0105 |     - |     - |      88 B |
|  PreComputeUnionConstructor |   Compiled |   7.737 ns |  0.0595 ns | 0.0497 ns | 0.0048 |     - |     - |      40 B |
|                             |            |            |            |           |        |       |       |           |
|       PreComputeUnionReader | Reflection | 318.424 ns |  4.2661 ns | 3.9905 ns | 0.0086 |     - |     - |      72 B |
|       PreComputeUnionReader |   Compiled |  13.706 ns |  0.3290 ns | 0.4822 ns | 0.0086 |     - |     - |      72 B |
|                             |            |            |            |           |        |       |       |           |
|    PreComputeUnionTagReader | Reflection | 373.670 ns |  2.4971 ns | 2.3358 ns | 0.0029 |     - |     - |      24 B |
|    PreComputeUnionTagReader |   Compiled |   6.567 ns |  0.1600 ns | 0.2136 ns |      - |     - |     - |         - |