@cartermp 2020-07-26
In this case I'm referring to a loop like this: https://github.com/dotnet/fsharp/pull/9784/files#diff-54378f53a8612b2adf50a6efb115ba5aR134

The compiler rewrites this to use an enumerator no matter which kind of `for` loop you use, see:

[current impl](https://sharplab.io/#v2:DYLgZgzgNAJiDUAfA9gBwKYDsAEBlAnhAC7oC2AsAFBpZ6EmkB0ASumMOgMZECWymVGjgLEyjADI9MAR0YBRAB6oATuggQ+mCFSoci2TslKoeHAKqZNAYQCGEdFf7FlAV27JlAMReZO2ABSk6EQAFsgwAJKYYMgg2ACywWGR0cgAlNgAvFTYudh62EGh4QAKNso2pBBZhUnhUTGMAOLBZRVF6MrV/mk5eQWo5ZUAgsoV+DWKKmoa/IxtlcGdAUT4GMhgADzIAEYAVgDaALoAfFDYAESD7Re9lH25BehKylkPeblTquqaEpU7MBsm28vm2+2O512exOJwC7w+Hy+M1+jkwADdOvp/PCEbikT85rZgMA4ZRceTyUVkg1kFAcRTcQd6QyKTFXtdKlEYM9sFJsAAGbCMRi1YowBZVCRYADmoWwAFpsABGbAwZDMlksgY1KmlIaSg4c0hc55HHRkzWWxEvZGE/gY5RYjVWln42aYRijcbDTicGYBI1emz4c5ulFOIg2TD6I0mhRpOkWl3J7Coeb6pbKAAqa3QzpTuQyIFhYf4+YZZqTVoT5Y+q3WWyhJxrVYpgbGweZdw+zxUjEcxlM6ACaSAA===)
[for x = 0 to loop](https://sharplab.io/#v2:DYLgZgzgNAJiDUAfA9gBwKYDsAEBlAnhAC7oC2AsAFBpZ6EmkB0ASumMOgMZECWymVGjgLEyjADI9MAR0YBRAB6oATuggQ+mCFSoci2TslKoeHAKqZNAYQCGEdFf7FlAV27JlAMReZO2ABSk6EQAFsgwAJKYYMgg2ACywWGR0cgAlNgAvFTYudh62EGh4QAKNso2pBBZhUnhUTGMAOLBZRVF6MrV/mk5eQWo5ZUAgsoV+DWKKmoa/IxtlcGdAUT4GMhgADzIAEYAVgDaALoAfFDYAESD7Re9lH25BehKylkPeblTquqaEpU7MBsm28vm2+2O512exOJwC7w+Hy+M1+jkwADdOvp/PCEbikT85rZgMA4ZRceTyUVkg1kFAcRTcQd6QyKTFXtdKlEYM8agAGbBEZC1YowBZVCRYADmoWwAFpsABGbAwZDMlksgY1KmlIbig4c0hc55HHRk9XmxEvZGE/gY5RYtUWln42aYRijcbDTicGYBA0emz4c4ulFOIg2TD6A1GhRpOlmp2J7Coea6pbKAAqa3QjqTuQyIFhIf4uYZJoTFrjpY+q3WWyhJyrFYp/rGgeZdw+zxUjEcxlM6ACaSAA==)

Rewriting it to a while loop and manually bumping the indexer will remove that [like so](https://sharplab.io/#v2:DYLgZgzgNAJiDUAfA9gBwKYDsAEBlAnhAC7oC2AsAFBpZ6EmkB0ASumMOgMZECWymVGjgLEyjADI9MAR0YBRAB6oATuggQ+mCFSoci2TslKoeHAKqZNAYQCGEdFf7FlAV27JlAMReZO2ABSk6EQAFsgwAJKYYMgg2ACywWGR0cgAlNgAvFTYudh62EGh4QAKNso2pBBZhUnhUTGMAOLBZRVF6MrV/mk5eQWo5ZUAgsoV+DWKKmoa/IxtlcGdAUT4GMhgADzIAEYAVgDaALoAfFDYAESD7Re9lH25BehKylkPeblTquqaEpU7MBsm28vm2+2O512exOJwC7w+Hy+M1+jkwADdOvp/PCEbikT85rZgMA4ZRceTyUVkg1kFAcRTcQd6QyKQVSC4iDYdhxsNdKlEYM8agAGZks3EAdxCpnQvKGpAFQs2tWKMAWVQkWAA5qFsDBkGLxazgryalTSvKIIwDnyFZhBQojjoyUbXZ8XsjCfwMcosYa3RT8bNMIxRuNhpxODMArawzZ8OcgyinJzMPpbYqFGk6S6AwHUPN5UtlAAVNbof157AZECwpP8St5xv5+WZ7CbAC0cvabfg2AAjM3cU7c67s0PsKt1lsoSdx6OKbGxvHmXcPs8VIxHMYZQE0kA=) but it's unlikely to matter much from a perf standpoint. Just kind of weird that trivially changing the `for` loop had no effect when it usually does.