@kerams 2020-07-26
Yeah, I'd be all for altering the loop were this happening inside of the delegate. Don't think it matters at all during prep stage.