@abelbraaksma 2020-07-26
In hot paths it matters for ints, it's about 4x slower. That  may not matter much here, though. There's an issue open to better optimize this, the example from @cartermp will help there.