@KevinRansom 2020-07-27
This func is compileUnionCaseConstructorFunc  ... it should not happen frequently.