@Daniel-Svensson 2020-07-28
@kerams if you want to further improve performance by 6-7 (on netcore 3.1) times for the "get all properties" case you can modify the expression to initialize the array in reverse order to avoid array bounds check. 

I have a method added to your benchmark from the previous PR and results laying around at home that I can post next week if it is of interest. 
But it might be regarded as a micro optimization and not too relevant here since one might expect  Expression.InitArray to be as fast as possible.