@kerams 2020-07-28
Are you saying 

```csharp
var a = int[5];
a[0]= 0;
a[0]= 0;
a[0]= 0;
```