@Daniel-Svensson 2020-07-28
@kerams actually I belive I must have a bug in my fsharp (cannot check it this week), it was a quick hack late before bedtime. 
If the length is unknown to the jit the reverse order is much faster but thinking more about it i would not expect such large improvements but only double digit percentage. There are many such "hacks" in the runtime